services/service_frontend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Bootstrap
PyJWT
Werkzeug
pytest
hypothesis
requests